Move `get_tdm_field_func_psi` to core
